Use dignified.js
