+++ .travis.yml
@@ -1,13 +1,15 @@
 sudo: false
 language: node_js
 node_js:
-  - "4.0"
+  - 4
+  - 5
 
 # Make sure we have new NPM.
 before_install:
   - npm install -g npm
 
 script:
+  - npm run lint
   - npm test
 
 addons:
