+++ test/browser.js
@@ -1,9 +1,9 @@
-'use strict'
 /* eslint-env mocha */
+'use strict'
 
 const expect = require('chai').expect
-const WSlibp2p = require('../../src')
 const multiaddr = require('multiaddr')
+const WSlibp2p = require('../src')
 
 describe('libp2p-websockets', function () {
   this.timeout(10000)
